backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
PyYAML>=6.0.1
pandas>=2.2.0
openpyxl>=3.1.0
python-multipart>=0.0.9
